# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "跨平台社交应用 API 服务",
        "title": "Social App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/calls/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的音视频通话房间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通话"
                ],
                "summary": "创建通话",
                "parameters": [
                    {
                        "description": "通话信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateCallInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/calls/ice-servers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取WebRTC连接所需的ICE服务器列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通话"
                ],
                "summary": "获取ICE服务器配置",
                "responses": {
                    "200": {
                        "description": "ICE服务器列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/calls/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "加入已有的通话房间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通话"
                ],
                "summary": "加入通话",
                "parameters": [
                    {
                        "description": "房间信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.JoinCallInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "加入成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "房间不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/calls/signal": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "处理WebRTC信令交换",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通话"
                ],
                "summary": "WebRTC信令",
                "parameters": [
                    {
                        "description": "信令数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SignalInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "处理成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/calls/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前所有通话房间的统计信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通话"
                ],
                "summary": "获取通话统计",
                "responses": {
                    "200": {
                        "description": "统计信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/calls/{room_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "离开通话房间",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通话"
                ],
                "summary": "离开通话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "离开成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.CreateCallInput": {
            "type": "object",
            "required": [
                "conversation_id",
                "type"
            ],
            "properties": {
                "conversation_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/webrtc.CallType"
                }
            }
        },
        "handler.JoinCallInput": {
            "type": "object",
            "required": [
                "room_id"
            ],
            "properties": {
                "room_id": {
                    "type": "string"
                }
            }
        },
        "handler.SignalInput": {
            "type": "object"
        },
...
        "webrtc.CallType": {
            "type": "string",
            "enum": [
                "audio",
                "video"
            ],
            "x-enum-varnames": [
                "CallTypeAudio",
                "CallTypeVideo"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}